backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.29.0",
    "boto3>=1.42.30",
    "fastapi>=0.128.0",
    "jinja2>=3.1.0",
    "bcrypt>=4.0.0,<4.1.0",
    "passlib>=1.7.4",
    "pydantic[email]>=2.0.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.21",
    "redis>=5.0.0",
    "slowapi>=0.1.9",
    "sqlalchemy[asyncio]>=2.0.0",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.13",
]

[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "double"
